refactor(auth): move reset password mutation to hooks directory
Relocate the reset password mutation hook from api directory to feature-specific hooks directory to better organize auth-related logic and improve maintainability
