+++ frontend/src/api/user/auth/mutations.ts
@@ -18,21 +18,7 @@ interface ArtistRequestPayload {
 
 
 
-export const useResetPasswordMutation = () => {
-  const navigate = useNavigate();
-  return useMutation({
-    mutationFn: (credentials: { token: string; password: string }) =>
-      apiClient.patch("/api/v1/auth/reset-password", credentials),
-    onSuccess: (res) => {
-      console.log("Password reset successful:", res.data);
-      toast.success("Password reset successful, Login now.");
-      navigate("/login");
-    },
-    onError: (error) => {
-      console.error("Password reset failed:", error);
-    },
-  });
-};
+
 
 export const changePasswordMutation = () => {
   return useMutation({
