+++ frontend/src/features/user/hooks/auth/useResetPasswordMutation.ts
@@ -1,0 +1,20 @@
+import { useMutation } from "@tanstack/react-query";
+import { useNavigate } from "react-router-dom";
+import apiClient from "../../../../api/axios";
+import toast from "react-hot-toast";
+
+export const useResetPasswordMutation = () => {
+  const navigate = useNavigate();
+  return useMutation({
+    mutationFn: (credentials: { token: string; password: string }) =>
+      apiClient.patch("/api/v1/auth/reset-password", credentials),
+    onSuccess: (res) => {
+      console.log("Password reset successful:", res.data);
+      toast.success("Password reset successful, Login now.");
+      navigate("/login");
+    },
+    onError: (error) => {
+      console.error("Password reset failed:", error);
+    },
+  });
+};
